Refactor importing modules and types.
