+++ packages/@markuplint/ml-core/src/ml-dom/tokens/comment.ts
@@ -1,7 +1,7 @@
-import { IMLDOMComment } from '../types';
-import { MLASTComment } from '@markuplint/ml-ast';
+import type { IMLDOMComment } from '../types';
+import type { MLASTComment } from '@markuplint/ml-ast';
 import MLDOMNode from './node';
-import { RuleConfigValue } from '@markuplint/ml-config';
+import type { RuleConfigValue } from '@markuplint/ml-config';
 
 export default class MLDOMComment<T extends RuleConfigValue, O = null>
 	extends MLDOMNode<T, O, MLASTComment>
