+++ packages/@markuplint/ml-core/src/ml-dom/tokens/preprocessor-specific-block.ts
@@ -1,8 +1,8 @@
-import { AnonymousNode, Document } from '../';
-import { IMLDOMPreprocessorSpecificBlock } from '../types';
-import { MLASTPreprocessorSpecificBlock } from '@markuplint/ml-ast';
-import { MLDOMNode } from './';
-import { RuleConfigValue } from '@markuplint/ml-config';
+import type { AnonymousNode, Document } from '../';
+import type { IMLDOMPreprocessorSpecificBlock } from '../types';
+import type { MLASTPreprocessorSpecificBlock } from '@markuplint/ml-ast';
+import MLDOMNode from './node';
+import type { RuleConfigValue } from '@markuplint/ml-config';
 
 export default class MLDOMPreprocessorSpecificBlock<T extends RuleConfigValue, O = null>
 	extends MLDOMNode<T, O, MLASTPreprocessorSpecificBlock>
